+++ src/const/names.js
@@ -1,0 +1,4 @@
+const NAMES = {
+  TITLE:"Dimbula",
+}
+export default Object.freeze(NAMES)
