+++ src/views/sign/ForgetPw.js
@@ -3,12 +3,12 @@ import React,{ useState } from "react"
 import { makeStyles } from '@material-ui/core/styles'
 import { 
   Box,
-  Container,
   Button,
   Divider,
   TextField,
 } from '@material-ui/core'
-import { signStyle } from './userStyle'
+import { signStyle } from './signStyle'
+import SignLayout from "./SignLayout"
 
 import firebase from "firebase/app"
 import "firebase/auth";
@@ -36,54 +36,47 @@ const ForgetPw = () => {
   }
 
   return(
-    <React.Fragment>
-      <Box my={4}>
-        <h1 className={classes.title} >Dimbula</h1>
+    <SignLayout>
+      <Box my={3}>
+        <h2 className={classes.sub_title}>Forget Password</h2>
       </Box>
 
-      <Container className={classes.signin_container} maxWidth="sm" >
-
-        <Box my={3}>
-          <h2 className={classes.sub_title}>Forget Password</h2>
-        </Box>
-
-        <Box className={classes.signin_form} autoComplete="off">
-            <TextField
-              required
+      <Box className={classes.signin_form} autoComplete="off">
+          <TextField
+            required
+            fullWidth
+            name="email"
+            label="Mail Address"
+            variant="outlined"
+            margin="none"
+            defaultValue={pw}
+            onChange={handleInputChange}
+            />
+          <Box mt={2}>
+            <Button
               fullWidth
-              name="email"
-              label="Mail Address"
-              variant="outlined"
-              margin="none"
-              defaultValue={pw}
-              onChange={handleInputChange}
-              />
-            <Box mt={2}>
-              <Button
-                fullWidth
-                variant="contained"
-                color="primary"
-                onClick={submit}
-                >
-                Reset Password
-              </Button>
-            </Box>
-        </Box>
+              variant="contained"
+              color="primary"
+              onClick={submit}
+              >
+              Reset Password
+            </Button>
+          </Box>
+      </Box>
 
-        <Box my={3}>
-          <Divider variant="middle" />
-        </Box>
-        
-        <Box textAlign={"right"} mb={4}>
-          <SignUpLink component="p" underline="always" />
-          <br/>
-          <SignInLink component="p" underline="always" />
-          <br/>
-          <ResendMailLink component="p" underline="always" />
-        </Box>
+      <Box my={3}>
+        <Divider variant="middle" />
+      </Box>
+      
+      <Box textAlign={"right"} mb={4}>
+        <SignUpLink component="p" underline="always" />
+        <br/>
+        <SignInLink component="p" underline="always" />
+        <br/>
+        <ResendMailLink component="p" underline="always" />
+      </Box>
 
-      </Container>
-    </React.Fragment>
+    </SignLayout>
   )
 
 }
