+++ src/views/sign/ResendEmail.js
@@ -4,12 +4,12 @@ import { useHistory } from "react-router-dom"
 import { makeStyles } from '@material-ui/core/styles'
 import { 
   Box,
-  Container,
   Button,
   Divider,
   TextField,
 } from '@material-ui/core'
-import { signStyle } from './userStyle'
+import { signStyle } from './signStyle'
+import SignLayout from "./SignLayout"
 
 import firebase from "firebase/app"
 import "firebase/auth";
@@ -37,7 +37,6 @@ const SignIn = () => {
   }
 
   const submit = () => {
-    // Todo implement try/catch
     firebase.auth().signInWithEmailAndPassword(
       info.email,
       info.password
@@ -58,66 +57,59 @@ const SignIn = () => {
   }
 
   return(
-    <React.Fragment>
-      <Box my={4}>
-        <h1 className={classes.title} >Dimbula</h1>
+    <SignLayout>
+      <Box my={3}>
+        <h2 className={classes.sub_title}>Re-Send Email</h2>
       </Box>
 
-      <Container className={classes.signin_container} maxWidth="sm" >
-
-        <Box my={3}>
-          <h2 className={classes.sub_title}>Re-Send Email</h2>
-        </Box>
-
-        <Box className={classes.signin_form} autoComplete="off">
-            <TextField
-              required
-              fullWidth
-              name="email"
-              label="Mail Address"
-              variant="outlined"
-              margin="none"
-              defaultValue={info.email}
-              onChange={handleInputChange}
-              />
-            <TextField
-              required
-              fullWidth
-              name="password"
-              label="Password"
-              type="password"
-              autoComplete="off"
-              variant="outlined"
-              margin="normal"
-              value={info.password}
-              onChange={handleInputChange}
+      <Box className={classes.signin_form} autoComplete="off">
+          <TextField
+            required
+            fullWidth
+            name="email"
+            label="Mail Address"
+            variant="outlined"
+            margin="none"
+            defaultValue={info.email}
+            onChange={handleInputChange}
             />
-            <Box mt={2}>
-              <Button
-                fullWidth
-                variant="contained"
-                color="primary"
-                onClick={submit}
-                >
-                Re-send
-              </Button>
-            </Box>
-        </Box>
+          <TextField
+            required
+            fullWidth
+            name="password"
+            label="Password"
+            type="password"
+            autoComplete="off"
+            variant="outlined"
+            margin="normal"
+            value={info.password}
+            onChange={handleInputChange}
+          />
+          <Box mt={2}>
+            <Button
+              fullWidth
+              variant="contained"
+              color="primary"
+              onClick={submit}
+              >
+              Re-send
+            </Button>
+          </Box>
+      </Box>
 
-        <Box my={3}>
-          <Divider variant="middle" />
-        </Box>
-        
-        <Box textAlign={"right"} mb={4}>
-          <SignUpLink component="p" underline="always" />
-          <br/>
-          <SignInLink component="p" underline="always" />
-          <br/>
-          <ForgetPwLink component="p" underline="always" />
-        </Box>
+      <Box my={3}>
+        <Divider variant="middle" />
+      </Box>
+      
+      <Box textAlign={"right"} mb={4}>
+        <SignUpLink component="p" underline="always" />
+        <br/>
+        <SignInLink component="p" underline="always" />
+        <br/>
+        <ForgetPwLink component="p" underline="always" />
+      </Box>
 
-      </Container>
-    </React.Fragment>
+    </SignLayout>
   )
 
 }
